\name{functionsForForeach-package}
\alias{functionsForForeach-package}
\alias{functionsForForeach}
\docType{package}
\title{\packageTitle{functionsForForeach}}
\description{\packageDescription{functionsForForeach}}
\section{Package Content}{\packageIndices{functionsForForeach}}
\author{\packageAuthor{functionsForForeach}}
\section{Maintainer}{\packageMaintainer{functionsForForeach}}
\keyword{package}
